Added mathjs as a dependency, and finished functionality
